:hammer: Remove endpoint package and refactor code
