+++ composeApp/src/commonMain/kotlin/com/crosspaste/app/EndpointInfoFactory.kt
@@ -1,6 +1,7 @@
-package com.crosspaste.endpoint
+package com.crosspaste.app
 
 import com.crosspaste.dao.sync.HostInfo
+import com.crosspaste.dto.sync.EndpointInfo
 
 interface EndpointInfoFactory {
