+++ composeApp/src/desktopMain/kotlin/com/crosspaste/net/DesktopSyncInfoFactory.kt
@@ -1,9 +1,9 @@
 package com.crosspaste.net
 
 import com.crosspaste.app.AppInfo
+import com.crosspaste.app.EndpointInfoFactory
 import com.crosspaste.dao.sync.HostInfo
 import com.crosspaste.dto.sync.SyncInfo
-import com.crosspaste.endpoint.EndpointInfoFactory
 
 class DesktopSyncInfoFactory(
     val appInfo: AppInfo,
